chore: pipe through root package json to lockfile parsing (#4437)
### Description

In order to properly support prune support for the [yarn `resolutions`
field](https://yarnpkg.com/configuration/manifest/#resolutions), we need
to have access to the `resolutions` field when we construct the lockfile
object. This PR simply adds the root package json as an argument to all
lockfile parsing methods.

### Testing Instructions

Existing unit and integration tests pass.