port(turborepo): New Turbopath (#4439)
### Description

This is meant as a very very simple port of turbopath to Rust. It's
basically the Go code but the Rust version.

The goal here is to make something that works, and that we can extend
with a little ChatGPT and copilot help.

### Testing Instructions

There are doc tests on relevant functions and some regular tests too.