ci: Stop building Go code twice (#4501)
### Description

When we use the makefile, we build the Go code, then build the Rust
code, but the Rust code in turn builds the Go code again.

This PR stops that by removing `go-turbo` as a dependency of `turbo` in
the Makefile. Instead we rely on the build.rs script to kick off
compilation of the Go code.

### Testing Instructions

Existing CI should test it. And hopefully be faster.