Daemon server rust port (#4306)
### Description

This ports the daemon server to rust, leaving only run an prune in
go-land. This needs some polish but I would like to get some feedback
now. I have also added a flag for running the grpc server over http
instead, which is useful for debugging. It has support for reflection,
so can be used via postman or grpcurl.

For code reviews, it should be ok to go in commit-order.

Outstanding parts:

- [x] incomplete tests in glob watch
- [x] error handling for watch errors
- [ ] logging / log rotation

 Out of scope (for now):

- startup performance pass

### Testing Instructions

This is a drop-in replacement for the go server, so `turbo run` in any
repo should work identically.