feat: Use Rust Berry lockfile impl (#4684)
### Description

Built on top of #4629

Overview of changes:
- ~~Change `subgraph` so we calculate workspace's closure in parallel in
a similar manner to Go~~ This had to get reverted, it appears that rayon
was causing [compilation issues on
linux](https://github.com/vercel/turbo/actions/runs/4791974785/jobs/8522988391#step:4:648)
 - `npm_subgraph` -> `subgraph` makes this FFI call generic
- Addition of an optional `resolutions` map to the `subgraph` and
`transitive_closure`, this is only used by Berry. In the ideal world
this would be all of the root `package.json`, but serializing all of
that via protobuf is overkill since berry is the only lockfile that
needs that info and we only use that field.
- Addition of `patches` FFI function, this wasn't needed before as NPM
doesn't have patch files encoded in the lockfile
- Addition of `global_changes` FFI function, previously for NPM we were
just doing this in Go since there were just two fields we needed to
check and parsing JSON is cheap.
- Switches lockfile usage of berry from the Go implementation to the
Rust one introduced in #4589

Reviewer notes:
- The passing of the optional resolutions map is very icky and one off,
but this is the only lockfile the requires additional information to
properly parse. Once the package graph is squarely in Rust land this
should go back to being more generic.

This will finally close out #2791

### Testing Instructions
Added a new integration test that uses resolutions. This integration
test invokes the new `patches` FFI call.

For a full test of the new resolution feature, clone and follow the
instructions in the
[repro](https://github.com/erj826/turbo-prune-bug-repro-11-21)

The `lockfile_aware_caching` integration tests all hit the new
`global_changes` FFI callsite and provide test coverage.

---------

Co-authored-by: Alexander Lyon <arlyon@me.com>