refactor(turborepo): cache signature authentication (#4686)
### Description

Refactors the cache signature authentication code to no longer rely on
Go implementation details.

This requires a cache bump as it is a **breaking** change to cache
signatures.

### Testing Instructions

Existing cache signature authentication now calls the Rust
implementation and confirms that it produces the same tag. There are
also new tests that randomly generate test cases and confirm that the
Rust and Go implementation agree on the tag output.