fix: allow for workplace descriptors without a protocol (#4755)
### Description

Fixes the issue found by @quinnturner in
[#2891](https://github.com/vercel/turbo/issues/2791#issuecomment-1526477128).

Crux of the issue was we were being too restrictive when tracking
dependency protocols to properly infer them for later construction. We
now allow for at most two descriptors in the protocol resolver: one
without a protocol and one with a protocol. Another change was no longer
adding all descriptors for a workspace and instead only including
descriptors that come from other workspaces.

### Testing Instructions

New testing fixture contains a lockfile where one workspace depends on
another via `*` and the other depends on it via `workspace:*`.