Fix TSConfig extends pointing to node module (#4772)
In #4754, I screwed up the resolution logic for configs that reference
another config inside a node module. In it, I incorrectly joined
`/tsconfig.json` before performing any lookup, because I thought the
joining was happening in the `combinePaths` in
https://github.com/microsoft/TypeScript/blob/611a912d/src/compiler/commandLineParser.ts#L3315.
But, that combine has no real affect, it immediately gets stripped in a
call to `getDirectoryPath` in
https://github.com/microsoft/TypeScript/blob/611a912d/src/compiler/moduleNameResolver.ts#LL1703C93-L1703.

Instead, it performs a basic `node_module` lookup using the input module
name, and even tries to load the file referenced at that location (eg,
it's a `extends: 'foo/tsconfig.json'`). If that fails, then it combines
with an implied `/tsconfig` the same way a `/index` is appended when
doing `require('./directory')`.

Fixes WEB-974