fix: no longer include peer dependencies in transitive closures (#4824)
### Description

Fixes #4757

Underlying issue was I was including `peerDependencies` in the
dependency crawl. This resulted in an error since we expect to only be
looking at external dependencies in the dependency crawl.

### Testing Instructions

Added a failing unit test where there's a workspace that's used as a
peer dependency that conflicts with a package in the registry.