Partial revert of #4820. (#4938)
This is a partial revert of 78aa3339bc21090c5c6c74f45bde776292ac8196 in
order to correct non-`git` fallback behavior.

1. It does not revert changes made inside of Rust `turbopath` as those
do not apply cleanly; but they are also not relevant to the issue
introduced by #4820.
2. The changes to `Cargo.lock` are the consequences of the new
dependency graph.

Tested for correctness locally, this restores behavior to what we saw
prior to the commit landing.

It's likely a 1-2 line change to correct behavior but I didn't attempt
to fix-forward. This will be straightforward to land again once we have
additional testing in place.

Co-authored-by: Nathan Hammond <Nathan Hammond>