Exclude Pipeline Definition from Global Hash (#4545)
The global hash does not need to include the pipeline so long as all
material elements to the pipeline are present in the task hash. After
#4529 that invariant became true.

This PR does two things:
- Removes `pipeline` from the global hash key inputs.
- Adjusts the dry run JSON schema to move `rootPipeline` from
`globalHashInputs` and instead attach it to a new root key, `pipeline`.
This is a possibly-breaking change so I incremented the
`runSummarySchemaVersion`.

🚨🚨🚨
This PR triggers a universal cache invalidation! The payoff is no longer
breaking the cache any time `turbo.json` is updated.

---------

Co-authored-by: Nathan Hammond <Nathan Hammond>