Support installing tip-of-canary next for benchmarks (#5156)
### Description

A small change to support installing a local
`next@file:path/to/next.js/packages/next` next package. This fixes an
issue where we run tip-of-canary `next-dev` vs the last published `next`
package, and allows us to benchmark breaking changes.

The
[`--install-links`](https://docs.npmjs.com/cli/v8/commands/npm-install#install-links)
option is a little oxymoronic. When set to true, it **doesn't** use
symlinks, so we need to set it to false to avoid copying a bunch of
files.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->