feat(lockfiles): add support pnpm lockfile version 6.1 (#5195)
### Description

Adds parse support for the new lockfile version added in
https://github.com/pnpm/pnpm/pull/6557 fixing #5117

The behavioral change is pretty small for us. The new information was
added to allow for `pnpm install --frozen-lockfile` to produce an
identical lockfile even if the `.npmrc` that was present during the
lockfile's creation is no longer present. This does lead to a better UX
for prune users as they will no longer need to copy `.npmrc` in their
`Dockerfile`.

### Testing Instructions

Existing unit tests pass and added new fixture with the settings field
that correctly parses.

---------

Co-authored-by: Chris Olszewski <Chris Olszewski>