Add codemod for glob syntax issues. (#5184)
### Description

We discovered a few classes of glob that are soon-to-be invalid with the
tightening on the syntax from `wax`. This codemod handles the most
common cases. Out of 720 unique globs encountered, only about 10 of them
were invalid, and of those there were 3 general classes of syntax
problem which are being addressed here, along with 3 just completely
invalid ones. This codemod should as a result fix around 80% of the
syntax issues.

### Testing Instructions

```
cd packages/turbo-codemod
pnpm test clean-globs
```

---------

Co-authored-by: Chris Olszewski <Chris Olszewski>
Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>