Prepare StaticAssetsContentSourceVc for basePath (#5210)
### Description

This is part 1 of WEB-993 `basePath` support. The static assets (things
served in the `/public` directory) can only serve requests that include
the `basePath`, so we'll need to pass that in. In order to not
completely invalidate the
[source](https://github.com/vercel/next.js/blob/2b1f0d9351610b04d01638efed19252ca81d0023/packages/next-swc/crates/next-dev/src/lib.rs#L379-L380)
function (and all the things that depend on cells created there, ie
everything) every time the user edits the `next.config.js`, we'll need
to accept a `StringVc`. This sets us up without breaking backwards
compat, making this multi-PR adventure a little easier.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->