ci(examples): make every examples test a workspace (#5011)
This commit parallelizes examples tests with turbo instead of Github Actions to:

- reduce the number of Github Workflow runs  triggered when an example
changes. This should be faster and easier to read in a PR (2 checks
instead of 16).
- more turbo dogfooding

- All examples tests will still run when any of the examples change,
because each example itself is not a workspace, so we cannot wire up
individual dependencies. 
- `turbo.json` configures `inputs` for caching, but we still using the
Github Actions config to to limit these runs to changes in `examples/`
and `examples-tests/**`. We could opt out of this and rely on turbo's
caching in the future.

- The `pnpm-gatsby` test was not part of the matrix before. Moving to
this setup revealed that it does not pass in CI (because of mismatched
node version). This test has been filtered out for now.