fix(turbo): print help message when turbo called with flags/env vars but no commands
Clap interprets env vars flags, and if flags are given it does not
error by default. This is expected behavior. Because we have code that
defaults no commands as run command (to make 'turbo build' work),
we need to handle printing the help message in this case, when no tasks
are given.