feat(ecma-analyzer): support super() call in class ctor (#5392)
### Description

Attempt to close WEB-1089.

PR tries to support import resolustion inside of `super` call, if given
args are referencing esm imports.

Since super() is sort of expr_call, I originally tried to reuse
JsValue::Call as-is but not successfully. Still that might be a simple
mistake and new addition to JsValue::Super may not be required.