Fix internal JS/TS references not being processed (#5457)
### Description

My recent change in #5397 added a way to create rules that only apply to
internal references. In the process, I also made it so the default rules
don't apply to internal references as well. However, we still need them
to apply to TS and JS files, as we use that with ReferenceType::Internal
in Next.js.

### Testing Instructions

Next.js CI