feat: port task graph validation (#5703)
### Description

Port of the task graph validation that checks the configuration of
persistent tasks.

### Testing Instructions

There are quite a few integration tests that test the various validation
errors. We will rely on those to catch any behavior changes from the Go
code rather than trying to convert those to unit tests. At the moment we
aren't running these tests with the Rust run, but that should come soon.

---------

Co-authored-by: Chris Olszewski <Chris Olszewski>