fix: copy root json to full directory (#5769)
### Description

Fixes #5768

We were copying the root `package.json` into `out` and not `out/full` as
intended if the `--docker` flag was used. This behavior would only
appear if the root `package.json` contained patches the needed to be
pruned. This PR gets us back to the Go behavior.

### Testing Instructions

Started also tracking the contents of `out/full` in an integration test


Closes TURBO-1227

---------

Co-authored-by: Chris Olszewski <Chris Olszewski>